backend/requirements.txt
asgiref==3.6.0
atomicwrites==1.4.1
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
colorama==0.4.6
cryptography==41.0.2
defusedxml==0.7.1
Django==4.2.3
django-debug-toolbar==2.2
django-filter==23.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-csv==2.1.1
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
exceptiongroup==1.1.0
Faker==12.0.1
flake8==6.0.0
idna==2.8
iniconfig==2.0.0
mccabe==0.7.0
mixer==7.1.2
more-itertools==9.0.0
mypy==0.991
mypy-extensions==0.4.3
oauthlib==3.2.2
packaging==23.0
pluggy==0.13.1
prettytable==3.5.0
py==1.11.0
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.7.0
pytest==5.3.5
pytest-django==3.8.0
pytest-pythonpath==0.7.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2022.7.1
requests==2.22.0
requests-oauthlib==1.3.1
six==1.14.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sorl-thumbnail==12.6.3
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.4.0
tzdata==2022.7
unicodecsv==0.14.1
urllib3==1.25.11
wcwidth==0.2.5
